+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
  - New `history.frozen` API allows temporarily freezing history collection.
  - New `lousy.widget.zoom` statusbar widget: shows current page zoom level.
+ - Added `log` signal, emitted whenever a message is logged.
 
 ### Fixed